Updating the API to account for audit table, and refactoring.
