+++ backend/api/views/constants.py
@@ -1,0 +1,14 @@
+AUDITEE_INFO_PREVIOUS_STEP_DATA_WE_NEED = [
+    "user_provided_organization_type",
+    "met_spending_threshold",
+    "is_usa_based",
+]
+
+AUDITEE_INFO_STEP_TWO_FIELDS = [
+    "auditee_fiscal_period_start",
+    "auditee_fiscal_period_end",
+]
+
+ACCESS_SUBMISSION_PREVIOUS_STEP_DATA_WE_NEED = (
+    AUDITEE_INFO_PREVIOUS_STEP_DATA_WE_NEED + AUDITEE_INFO_STEP_TWO_FIELDS
+)
